third_party/WebKit/LayoutTests/inspector/sass/test-ssp-breaking-edits.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/elements-test.js"></script>
<script src="../../http/tests/inspector/debugger-test.js"></script>
<script src="./sass-test.js"></script>
<link rel="stylesheet">
<script>
function insertStyleSheet()
{
    var link = document.querySelector("link");
    link.setAttribute("href", "resources/test-ssp-breaking-edits.css");
}

function test()
{
    var sourceCode;
    Runtime.experiments.enableForTest("liveSASS");
    InspectorTest.evaluateInPage("insertStyleSheet()", onStyleSheet);

    function onStyleSheet()
    {
        InspectorTest.waitForScriptSource("test-ssp-breaking-edits.scss", onSASS);
    }

    function onSASS(sc)
    {
        sourceCode = sc;
        InspectorTest.selectNodeAndWaitForStyles("inspected", step1);
    }

    function step1()
    {
        var treeElement = InspectorTest.getMatchedStylePropertyTreeItem("font-family");
        treeElement.startEditing(treeElement.valueElement);
        treeElement.valueElement.textContent = "\"Times";
        treeElement.kickFreeFlowStyleEditForTest();
        InspectorTest.waitForStyleApplied(step2.bind(null, treeElement));
    }

    function step2(treeElement)
    {
        treeElement.valueElement.textContent = "\"Times New Roman\"";
        treeElement.valueElement.dispatchEvent(InspectorTest.createKeyEvent("Enter"));
        InspectorTest.waitForStyleApplied(step3);
    }

    function step3()
    {
        InspectorTest.addResult(sourceCode.workingCopy());
        InspectorTest.completeTest();
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that incremental breaking change doesn't detach sourcemap.
</p>

<div id="inspected"></div>
</body>
</html>
