third_party/WebKit/LayoutTests/inspector/sass/test-ssp-breaking-edits-expected.txt

Tests that incremental breaking change doesn't detach sourcemap.

$color: red;
div {
    font-family: "Times New Roman";
    color: $color;
}

body {
    padding: 0;
}
